+++ Client/Utils/I40ClientUtils/format_conversion.py
@@ -1,4 +1,8 @@
+from io import BytesIO
 from html2text import HTML2Text
+from weasyprint import HTML as WP_HTML
+from pdf2docx import Converter as PDF2DOCX_Converter
+import base64
 
 def HTML_To_Markdown(Content: str) -> str:
     h = HTML2Text(bodywidth = 0)
@@ -8,3 +12,25 @@ def HTML_To_Markdown(Content: str) -> str:
     
     h.close()
     return content
+
+def HTML_To_PDF(Content: str, ReturnEncoded: bool = True) -> bytes | str:
+    html = WP_HTML(string = Content)
+    data = html.write_pdf()
+
+    return base64.b64encode(data).decode("utf-8") if (ReturnEncoded) else data
+
+def PDF_To_DOCX(Content: str | bytes, ReturnEncoded: bool = True) -> bytes | str:
+    if (isinstance(Content, str)):
+        content = base64.b64decode(Content)
+    else:
+        content = Content
+    
+    buffer = BytesIO(content)
+    converter = PDF2DOCX_Converter(stream = content)
+    converter.convert(buffer)
+
+    output = buffer.getvalue()
+    buffer.close()
+    converter.close()
+
+    return base64.b64encode(output).decode("utf-8") if (ReturnEncoded) else output